glibc.malloc.check: Wean away from malloc hooks
The malloc-check debugging feature is tightly integrated into glibc
malloc, so thanks to an idea from Florian Weimer, much of the malloc
implementation has been moved into libc_malloc_debug.so to support
malloc-check.  Due to this, glibc malloc and malloc-check can no
longer work together; they use altogether different (but identical)
structures for heap management.  This should not make a difference
though since the malloc check hook is not disabled anywhere.
malloc_set_state does, but it does so early enough and that behaviour
has been ported to the debug DSO.

The only changes to malloc.c in this change is to add #if IS_IN (libc)
to not compile some functions into the debug DSO, especially the
allocator function entry points.

The malloc check tunable is now in the debug DSO and has no effect
when the DSO is not preloaded.